Refactor: 선택된 탭 recoil로 하도록 수정
